example/openrouter/pyproject.toml
[tool.poetry]
name = "openrouter"
version = "0.1.0"
description = ""
authors = ["Alexander-Ploskin <ploskin0107@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
rebel-eval = {extras = ["deepeval"], version = "^1.1.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
